+++ frameworks/react/workspace-platform-starter/.gitignore
@@ -22,3 +22,7 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+
+# Runtime files copied to public folder
+public/openfin
+public/common
