+++ frameworks/react/workspace-platform-starter/package.json
@@ -4,9 +4,11 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
+    "prestart": "npm run build:openfin",
     "dev": "vite",
-    "start": "vite preview --port 8080",
-    "build": "tsc -b && vite build",
+    "start": "npm run dev",
+    "build": "tsc -b && npm run build:openfin && vite build",
+    "build:openfin": "rollup -c rollup.config.mjs",
     "lint": "eslint .",
     "preview": "vite preview --port 8080"
   },
@@ -24,6 +26,10 @@
     "@eslint/js": "^9.25.0",
     "@openfin/core": "41.102.1",
     "@openfin/node-adapter": "41.102.1",
+    "@rollup/plugin-commonjs": "^28.0.3",
+    "@rollup/plugin-node-resolve": "^16.0.1",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-typescript": "^12.1.2",
     "@types/react": "^19.1.2",
     "@types/react-dom": "^19.1.2",
     "@vitejs/plugin-react": "^4.4.1",
@@ -31,6 +37,7 @@
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "globals": "^16.0.0",
+    "rollup-plugin-copy": "^3.5.0",
     "typescript": "~5.8.3",
     "typescript-eslint": "^8.30.1",
     "vite": "^6.3.5",
