+++ frameworks/react/workspace-platform-starter/src/Provider.tsx
@@ -4,12 +4,12 @@ export function Provider() {
 	useOpenFin();
 
 	return (
-		<div>
+		<main>
 			<h3>OpenFin platform provider</h3>
 			<p>
 				The window would usually be hidden, you can make it hidden on startup by setting the
 				platform.autoShow flag to false in the manifest.fin.json
 			</p>
-		</div>
+		</main>
 	);
 }
