+++ frameworks/react/workspace-platform-starter/src/index.css
@@ -30,6 +30,10 @@ body {
   min-height: 100vh;
 }
 
+main {
+	padding: 2rem;
+}
+
 h1 {
   font-size: 3.2em;
   line-height: 1.1;
